# packages/contracts/src/homebrew_mlflow/contracts/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "Homebrew MLflow API",
    "version": "0.1.0"
  },
  "servers": [{"url": "/"}],
  "paths": {
...
    "/api/v1/auth/machine": {
      "post": {
        "operationId": "authenticateMachineCredential", "summary": "Exchange a scoped machine credential for a short-lived token",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MachineLoginRequest"}}}},
        "responses": {"200": {"description": "Short-lived scoped token", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccessTokenResponse"}}}}, "401": {"description": "Invalid machine credential", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}}
      }
    },
...
    "/api/v1/projects": {
      "get": {
        "operationId": "listProjects", "summary": "List the caller's Research Projects", "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "Research Projects", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/ProjectResponse"}}}}}}
      },
      "post": {
        "operationId": "createProject",
        "summary": "Create a Research Project and its seeded default repository",
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateProjectRequest"}}}},
        "responses": {
          "202": {"description": "Project and repository provisioning accepted", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectCreationResponse"}}}},
          "403": {"description": "Organization Admin required", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
          "409": {"description": "Project slug conflict", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}
        }
      }
    },
...
    "/api/v1/projects/{project_id}/memberships": {
      "get": {
        "operationId": "listProjectMemberships", "summary": "List canonical project membership", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Project membership", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/ProjectMembershipResponse"}}}}}}
      }
    },
...
    "/api/v1/projects/{project_id}/memberships/{principal_id}": {
      "put": {
        "operationId": "setProjectMembership", "summary": "Add or change a canonical project role", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string"}}, {"name": "principal_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SetProjectMembershipRequest"}}}},
        "responses": {"200": {"description": "Canonical membership queued for integration reconciliation", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectMembershipResponse"}}}}, "403": {"description": "Maintainer required", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}, "409": {"description": "Last Maintainer protection", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}}
      },
      "delete": {
        "operationId": "removeProjectMembership", "summary": "Remove canonical project membership", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string"}}, {"name": "principal_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"204": {"description": "Membership removed and reconciliation queued"}, "409": {"description": "Last Maintainer protection", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}}
      }
    },
    "/api/v1/projects/{project_id}/memberships/{principal_id}/recover-maintainer": {
      "post": {
        "operationId": "recoverProjectMaintainer", "summary": "Assign a human Maintainer through audited organization-Admin recovery", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string"}}, {"name": "principal_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Recovered Maintainer membership", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectMembershipResponse"}}}}, "403": {"description": "Organization Admin required", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}}
      }
    },
...
    "/api/v1/projects/{project_id}/machine-credentials": {
      "get": {
        "operationId": "listMachineCredentials", "summary": "List project machine credentials without secret digests", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Machine credentials", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/MachineCredentialSummary"}}}}}}
      },
      "post": {
        "operationId": "createMachineCredential", "summary": "Create a project-bound scoped machine Principal", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateMachineCredentialRequest"}}}},
        "responses": {"201": {"description": "One-time machine credential secret", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MachineCredentialResponse"}}}}, "403": {"description": "Maintainer required", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}}
      }
    },
    "/api/v1/machine-credentials/{credential_id}": {
      "delete": {
        "operationId": "revokeMachineCredential", "summary": "Revoke a project machine credential", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "credential_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Revoked machine credential", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MachineCredentialSummary"}}}}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}
    },
    "schemas": {
      "AccessTokenResponse": {
        "type": "object", "additionalProperties": false,
        "required": ["access_token", "token_type", "expires_in"],
        "properties": {"access_token": {"type": "string"}, "token_type": {"type": "string", "const": "Bearer"}, "expires_in": {"type": "integer", "const": 1200}}
      },
...
      "CreateMachineCredentialRequest": {
        "type": "object", "additionalProperties": false,
        "required": ["display_name", "role", "scopes"],
        "properties": {"display_name": {"type": "string", "maxLength": 200}, "role": {"type": "string", "enum": ["viewer", "contributor"]}, "scopes": {"type": "array", "minItems": 1, "uniqueItems": true, "items": {"type": "string", "enum": ["read", "track", "dvc_transfer", "publish"]}}}
      },
      "MachineCredentialResponse": {
        "type": "object", "additionalProperties": false,
        "required": ["credential_id", "principal_id", "project_id", "secret", "scopes", "expires_at"],
        "properties": {"credential_id": {"type": "string"}, "principal_id": {"type": "string"}, "project_id": {"type": "string"}, "secret": {"type": "string", "writeOnly": true}, "scopes": {"type": "array", "items": {"type": "string"}}, "expires_at": {"type": "string", "format": "date-time"}}
      },
      "MachineCredentialSummary": {
        "type": "object", "additionalProperties": false,
        "required": ["credential_id", "principal_id", "project_id", "scopes", "revoked", "expires_at"],
        "properties": {"credential_id": {"type": "string"}, "principal_id": {"type": "string"}, "project_id": {"type": "string"}, "scopes": {"type": "array", "items": {"type": "string"}}, "revoked": {"type": "boolean"}, "expires_at": {"type": "string", "format": "date-time"}}
      },
      "MachineLoginRequest": {
        "type": "object", "additionalProperties": false, "required": ["credential_id", "secret"],
        "properties": {"credential_id": {"type": "string"}, "secret": {"type": "string", "writeOnly": true}}
      },
...
      "CreateProjectRequest": {
        "type": "object", "additionalProperties": false,
        "required": ["organization_id", "name", "slug"],
        "properties": {
          "organization_id": {"type": "string", "pattern": "^org_[0-9A-HJKMNP-TV-Z]{26}$"},
          "name": {"type": "string", "minLength": 1, "maxLength": 200},
          "slug": {"type": "string", "minLength": 1, "maxLength": 100}
        }
      },
...
      "RepositoryResponse": {
        "type": "object", "additionalProperties": false,
        "required": ["id", "project_id", "name", "slug", "default_branch", "state", "web_url", "http_clone_url", "ssh_clone_url", "failure_code"],
        "properties": {
          "id": {"type": "string", "pattern": "^repo_[0-9A-HJKMNP-TV-Z]{26}$"},
          "project_id": {"type": "string", "pattern": "^pr_[0-9A-HJKMNP-TV-Z]{26}$"},
          "name": {"type": "string"}, "slug": {"type": "string"},
          "default_branch": {"type": "string"}, "state": {"type": "string", "enum": ["provisioning", "active", "failed", "archived"]},
          "web_url": {"type": ["string", "null"]}, "http_clone_url": {"type": ["string", "null"]},
          "ssh_clone_url": {"type": ["string", "null"]}, "failure_code": {"type": ["string", "null"]}
        }
      },
...
      "ProjectCreationResponse": {
        "type": "object", "additionalProperties": false,
        "required": ["id", "organization_id", "name", "slug", "default_repository"],
        "properties": {
          "id": {"type": "string", "pattern": "^pr_[0-9A-HJKMNP-TV-Z]{26}$"},
          "organization_id": {"type": "string", "pattern": "^org_[0-9A-HJKMNP-TV-Z]{26}$"},
          "name": {"type": "string"}, "slug": {"type": "string"},
          "default_repository": {"$ref": "#/components/schemas/RepositoryResponse"}
        }
      },
      "ProjectResponse": {
        "type": "object", "additionalProperties": false,
        "required": ["id", "organization_id", "name", "slug", "state", "created_at", "archived_at"],
        "properties": {"id": {"type": "string"}, "organization_id": {"type": "string"}, "name": {"type": "string"}, "slug": {"type": "string"}, "state": {"type": "string"}, "created_at": {"type": "string", "format": "date-time"}, "archived_at": {"type": ["string", "null"], "format": "date-time"}}
      },
...
      "SetProjectMembershipRequest": {
        "type": "object", "additionalProperties": false, "required": ["role"],
        "properties": {"role": {"type": "string", "enum": ["viewer", "contributor", "maintainer"]}}
      },
      "ProjectMembershipResponse": {
        "type": "object", "additionalProperties": false,
        "required": ["principal_id", "display_name", "principal_kind", "gitlab_username", "role", "created_at"],
        "properties": {
          "principal_id": {"type": "string"}, "display_name": {"type": "string"},
          "principal_kind": {"type": "string", "enum": ["human", "machine"]},
          "gitlab_username": {"type": ["string", "null"]},
          "role": {"type": "string", "enum": ["viewer", "contributor", "maintainer"]},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message", "request_id", "details", "retryable"],
        "properties": {
          "code": {"type": "string"},
          "message": {"type": "string"},
          "request_id": {"type": "string"},
          "details": {"type": "object"},
          "retryable": {"type": "boolean"}
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {"error": {"$ref": "#/components/schemas/Error"}}
      }
    }
  }
}